Added phpunits.

+++ migrations/Version20231113094709.php
@@ -14,20 +14,23 @@ final class Version20231113094709 extends AbstractMigration
 {
     public function up(Schema $schema): void
     {
-        $this->addSql('DROP SEQUENCE messenger_messages_id_seq CASCADE');
         $this->addSql('CREATE SEQUENCE hotel_id_seq INCREMENT BY 1 MINVALUE 1 START 1');
         $this->addSql('CREATE TABLE hotel (id INT NOT NULL, money_id INT DEFAULT NULL, search_id INT DEFAULT NULL, title VARCHAR(255) NOT NULL, url VARCHAR(1000) NOT NULL, image TEXT NOT NULL, address VARCHAR(255) NOT NULL, descriptions JSON NOT NULL, rate DOUBLE PRECISION DEFAULT NULL, source VARCHAR(255) NOT NULL, PRIMARY KEY(id))');
         $this->addSql('CREATE UNIQUE INDEX UNIQ_3535ED9BF29332C ON hotel (money_id)');
         $this->addSql('CREATE INDEX IDX_3535ED9650760A9 ON hotel (search_id)');
         $this->addSql('ALTER TABLE hotel ADD CONSTRAINT FK_3535ED9BF29332C FOREIGN KEY (money_id) REFERENCES money (id) ON DELETE CASCADE NOT DEFERRABLE INITIALLY IMMEDIATE');
         $this->addSql('ALTER TABLE hotel ADD CONSTRAINT FK_3535ED9650760A9 FOREIGN KEY (search_id) REFERENCES search (id) NOT DEFERRABLE INITIALLY IMMEDIATE');
-        $this->addSql('DROP TABLE messenger_messages');
         $this->addSql('ALTER TABLE search ADD from_at DATE NOT NULL');
         $this->addSql('ALTER TABLE search ADD to_at DATE NOT NULL');
         $this->addSql('ALTER TABLE search ADD adults INT NOT NULL');
         $this->addSql('ALTER TABLE search ADD children INT NOT NULL');
         $this->addSql('COMMENT ON COLUMN search.from_at IS \'(DC2Type:date_immutable)\'');
         $this->addSql('COMMENT ON COLUMN search.to_at IS \'(DC2Type:date_immutable)\'');
+
+        if ('test' !== $_ENV['APP_ENV']) {
+            $this->addSql('DROP SEQUENCE messenger_messages_id_seq CASCADE');
+            $this->addSql('DROP TABLE messenger_messages');
+        }
     }
 
     public function down(Schema $schema): void
